feat: finish wiring up the job to the script execution

+++ apps/web/db/migrations/1757010370062_initial_schema.ts
@@ -146,7 +146,7 @@ export async function up(db: Kysely<any>): Promise<void> {
     .addColumn("contract_id", "integer", (col) =>
       col.notNull(),
     )
-    .addColumn("price", "numeric", (col) => col.notNull())
+    .addColumn("price", "integer", (col) => col.notNull())
     .addColumn("start_date", "timestamptz")
     .addColumn("end_date", "timestamptz")
     .addColumn("created_at", "timestamptz", (col) =>
